+++ lib/country_picker_dialog.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
-import 'package:intl_phone_field/countries.dart';
-import 'package:intl_phone_field/helpers.dart';
+import 'countries.dart';
+import 'helpers.dart';
 
 class PickerDialogStyle {
   final Color? backgroundColor;