+++ lib/helpers.dart
@@ -1,4 +1,4 @@
-import 'package:intl_phone_field/countries.dart';
+import 'countries.dart';
 
 bool isNumeric(String s) => s.isNotEmpty && int.tryParse(s.replaceAll("+", "")) != null;
 
@@ -19,10 +19,11 @@ extension CountryExtensions on List<Country> {
     return where(
       (country) => isNumeric(search) || search.startsWith("+")
           ? country.dialCode.contains(search)
-          : removeDiacritics(country.name.replaceAll("+", "").toLowerCase()).contains(search) ||
+          : removeDiacritics(country.name.replaceAll("+", "").toLowerCase()).contains(search.toLowerCase()) ||
               country.nameTranslations.values
                   .any((element) => removeDiacritics(element.toLowerCase()).contains(search)) ||
-              country.code.toLowerCase().contains(search),
+              country.code.toLowerCase().contains(search.toLowerCase()) ||
+              country.currencyName.toLowerCase().contains(search.toLowerCase()),
     ).toList();
   }
 }